+++ CHANGELOG.md
@@ -4,11 +4,19 @@
 
 ### Added
 
+- Added `containerCornerRadius`, `equalButtonWidths`, and `minWidth` to `DefaultSwipeActionsView.Style` for additional swipe action style customization.
+- Added `swipeActionsViewStyle` to `ListEnvironment`. This allows a `SwipeActionsView.Style` to be set on the list environment when customizing the appearance of the swipe action view.
 - Added `containerCornerRadius`, `buttonSizing`, `minWidth`, and `maxWidthRatio` to `DefaultSwipeActionsView.Style` for additional swipe action style customization.
 
 ### Removed
 
+- Removed the `SwipeActionsView` associated type on `ItemContent`.
+- Removed the `ItemContentSwipeActionsView` protocol.
+
 ### Changed
+
+- Renamed `DefaultSwipeActionsView` to `SwipeActionsView`.
+- The type of the `ItemContent.swipeActionsStyle` protocol requirement is now `SwipeActionsView.Style?` (previously `SwipeActionsView.Style`). When an item returns `nil` for this property the style set on the list environment will be used instead.
 
 ### Misc